requirements.txt
streamlit
google-generativeai
pypdf 
pyyaml
python-dotenv